refactor(core): the navigator seam users touch moves out of internal - review finding #1
PollContext.getNavigatorView() and ParallelConsumerOptions.resourceAllocator
forced every user to import bz.stub.parallelconsumer.internal.navigator
types - an API-boundary leak cheap to fix before this surface ships and
breaking after. The user-facing seam moves to the public
bz.stub.parallelconsumer.navigator package: NavigatorView,
ResourceAllocator, ResourceContract, ResourceDeferral, plus the
transitive closure of their signatures (CapacityLease,
ConservationLedger) and StubResourceAllocator, the allocator users
actually construct. Engine wiring stays internal: NavigatorParticipant,
ParticipantBackedNavigatorView, NavigatorDecision/Reason.

NavigatorView.of(NavigatorParticipant, Clock) would have dragged an
internal type into a public signature - the exact defect class being
fixed - so the factory relocates to ParticipantBackedNavigatorView.of()
(internal), and that view's constructor narrows to private. The only
widenings are on that one internal class, for the cross-package inert()
and PCModule wiring seams.

All 11 moves recorded as git renames, correctly paired. The integration
test now imports one internal type instead of six (LOG_PREFIX, whitebox
log assertion). Scoped suites 106 green including the wall-clock IT and
the conventions ArchUnit checks; the dated plan's internal-package
citations left untouched per docs/citations.md.
