+++ parallel-consumer-core/src/main/java/bz/stub/parallelconsumer/navigator/ResourceDeferral.java
@@ -1,4 +1,4 @@
-package bz.stub.parallelconsumer.internal.navigator;
+package bz.stub.parallelconsumer.navigator;
 
 /*-
  * Copyright (C) 2026 Antony Stubbs and contributors
@@ -10,13 +10,14 @@ import java.time.Instant;
 import java.util.Optional;
 
 /**
- * One binding resource in a {@link NavigatorDecision} (U4, R9): the resource's name and when this member next
- * gains spendable credit against it. One instance per currently-blocking tagged resource - a record blocked on
- * several resources carries several of these, never a chosen one (R9's all-binding-predicates clause).
+ * One binding resource in a {@link bz.stub.parallelconsumer.internal.navigator.NavigatorDecision} (U4, R9):
+ * the resource's name and when this member next gains spendable credit against it. One instance per
+ * currently-blocking tagged resource - a record blocked on several resources carries several of these, never a
+ * chosen one (R9's all-binding-predicates clause).
  * <p>
  * {@link #getNextCreditAt()} is empty exactly when {@link ResourceAllocator#nextCreditAt} is: the resource's
- * policy mints nothing, so there is no time to name (mirrors {@link NavigatorParticipant#availableAt}'s own
- * emptiness case).
+ * policy mints nothing, so there is no time to name (mirrors
+ * {@link bz.stub.parallelconsumer.internal.navigator.NavigatorParticipant#availableAt}'s own emptiness case).
  */
 @Value
 public class ResourceDeferral {
