+++ parallel-consumer-core/src/test-integration/java/bz/stub/parallelconsumer/integrationTests/NavigatorRateShareTest.java
@@ -8,12 +8,12 @@ import bz.stub.parallelconsumer.ParallelConsumerOptions;
 import bz.stub.parallelconsumer.ParallelEoSStreamProcessor;
 import bz.stub.parallelconsumer.integrationTests.utils.KafkaClientUtils;
 import bz.stub.parallelconsumer.internal.PCModule;
-import bz.stub.parallelconsumer.internal.navigator.ConservationLedger;
+import bz.stub.parallelconsumer.navigator.ConservationLedger;
 import bz.stub.parallelconsumer.internal.navigator.NavigatorParticipant;
-import bz.stub.parallelconsumer.internal.navigator.NavigatorView;
-import bz.stub.parallelconsumer.internal.navigator.ResourceContract;
-import bz.stub.parallelconsumer.internal.navigator.ResourceDeferral;
-import bz.stub.parallelconsumer.internal.navigator.StubResourceAllocator;
+import bz.stub.parallelconsumer.navigator.NavigatorView;
+import bz.stub.parallelconsumer.navigator.ResourceContract;
+import bz.stub.parallelconsumer.navigator.ResourceDeferral;
+import bz.stub.parallelconsumer.navigator.StubResourceAllocator;
 import bz.stub.parallelconsumer.state.ProcessingShard;
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
