+++ parallel-consumer-core/src/test/java/bz/stub/parallelconsumer/internal/navigator/PCModuleResourceAllocatorAccessorTest.java
@@ -4,6 +4,7 @@ package bz.stub.parallelconsumer.internal.navigator;
  * Copyright (C) 2026 Antony Stubbs and contributors
  */
 
+import bz.stub.parallelconsumer.navigator.StubResourceAllocator;
 import bz.stub.parallelconsumer.ParallelConsumerOptions;
 import bz.stub.parallelconsumer.internal.PCModule;
 import bz.stub.parallelconsumer.internal.utils.LongPollingMockConsumer;
