+++ parallel-consumer-core/src/test/java/bz/stub/parallelconsumer/state/NavigatorAttributionTest.java
@@ -8,9 +8,9 @@ import bz.stub.parallelconsumer.ParallelConsumerOptions;
 import bz.stub.parallelconsumer.internal.EpochAndRecordsMap;
 import bz.stub.parallelconsumer.internal.PCModuleTestEnv;
 import bz.stub.parallelconsumer.internal.navigator.NavigatorDecisionReason;
-import bz.stub.parallelconsumer.internal.navigator.ResourceAllocator;
-import bz.stub.parallelconsumer.internal.navigator.ResourceContract;
-import bz.stub.parallelconsumer.internal.navigator.StubResourceAllocator;
+import bz.stub.parallelconsumer.navigator.ResourceAllocator;
+import bz.stub.parallelconsumer.navigator.ResourceContract;
+import bz.stub.parallelconsumer.navigator.StubResourceAllocator;
 import bz.stub.parallelconsumer.metrics.PCMetricsDef;
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
