+++ parallel-consumer-core/src/test/java/bz/stub/parallelconsumer/state/NavigatorSelectionTest.java
@@ -7,10 +7,10 @@ package bz.stub.parallelconsumer.state;
 import bz.stub.parallelconsumer.ParallelConsumerOptions;
 import bz.stub.parallelconsumer.internal.EpochAndRecordsMap;
 import bz.stub.parallelconsumer.internal.PCModuleTestEnv;
-import bz.stub.parallelconsumer.internal.navigator.ConservationLedger;
-import bz.stub.parallelconsumer.internal.navigator.ResourceAllocator;
-import bz.stub.parallelconsumer.internal.navigator.ResourceContract;
-import bz.stub.parallelconsumer.internal.navigator.StubResourceAllocator;
+import bz.stub.parallelconsumer.navigator.ConservationLedger;
+import bz.stub.parallelconsumer.navigator.ResourceAllocator;
+import bz.stub.parallelconsumer.navigator.ResourceContract;
+import bz.stub.parallelconsumer.navigator.StubResourceAllocator;
 import bz.stub.parallelconsumer.state.WorkContainer.ExecutionState;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
